requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml
python-dotenv==1.0.0
pymysql==1.1.0
fake-useragent==1.4.0
Jinja2>=3.1.2
python-telegram-bot==20.7